Split url tests out into separate classes
