+++ src/Tests/Tests/XPack/Slm/DeleteLifecycle/DeleteLifecycleUrlTests.cs
@@ -1,0 +1,18 @@
+﻿using System.Threading.Tasks;
+using Elastic.Xunit.XunitPlumbing;
+using Nest;
+using Tests.Framework.EndpointTests;
+using static Tests.Framework.EndpointTests.UrlTester;
+
+namespace Tests.XPack.Slm.DeleteLifecycle
+{
+	public class DeleteLifecycleUrlTests : UrlTestsBase
+	{
+		[U] public override async Task Urls() =>
+			await DELETE("/_slm/policy/policy_id")
+				.Fluent(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycle("policy_id"))
+				.Request(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycle(new DeleteSnapshotLifecycleRequest("policy_id")))
+				.FluentAsync(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycleAsync("policy_id"))
+				.RequestAsync(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycleAsync(new DeleteSnapshotLifecycleRequest("policy_id")));
+	}
+}
