+++ src/Tests/Tests/XPack/Slm/GetLifecycle/GetLifecycleUrlTests.cs
@@ -4,19 +4,12 @@ using Nest;
 using Tests.Framework.EndpointTests;
 using static Tests.Framework.EndpointTests.UrlTester;
 
-namespace Tests.XPack.Slm
+namespace Tests.XPack.Slm.GetLifecycle
 {
-	public class SlmUrlTests : UrlTestsBase
+	public class GetLifecycleUrlTests : UrlTestsBase
 	{
 		[U] public override async Task Urls()
 		{
-			await PUT("/_slm/policy/policy_id")
-					.Fluent(c => c.SnapshotLifecycleManagement.PutSnapshotLifecycle("policy_id"))
-					.Request(c => c.SnapshotLifecycleManagement.PutSnapshotLifecycle(new PutSnapshotLifecycleRequest("policy_id")))
-					.FluentAsync(c => c.SnapshotLifecycleManagement.PutSnapshotLifecycleAsync("policy_id"))
-					.RequestAsync(c => c.SnapshotLifecycleManagement.PutSnapshotLifecycleAsync(new PutSnapshotLifecycleRequest("policy_id")))
-				;
-
 			await GET("/_slm/policy/policy_id")
 					.Fluent(c => c.SnapshotLifecycleManagement.GetSnapshotLifecycle(g => g.PolicyId("policy_id")))
 					.Request(c => c.SnapshotLifecycleManagement.GetSnapshotLifecycle(new GetSnapshotLifecycleRequest("policy_id")))
@@ -36,20 +29,6 @@ namespace Tests.XPack.Slm
 					.Request(c => c.SnapshotLifecycleManagement.GetSnapshotLifecycle(new GetSnapshotLifecycleRequest()))
 					.FluentAsync(c => c.SnapshotLifecycleManagement.GetSnapshotLifecycleAsync())
 					.RequestAsync(c => c.SnapshotLifecycleManagement.GetSnapshotLifecycleAsync(new GetSnapshotLifecycleRequest()))
-				;
-
-			await PUT("/_slm/policy/policy_id/_execute")
-					.Fluent(c => c.SnapshotLifecycleManagement.ExecuteSnapshotLifecycle("policy_id"))
-					.Request(c => c.SnapshotLifecycleManagement.ExecuteSnapshotLifecycle(new ExecuteSnapshotLifecycleRequest("policy_id")))
-					.FluentAsync(c => c.SnapshotLifecycleManagement.ExecuteSnapshotLifecycleAsync("policy_id"))
-					.RequestAsync(c => c.SnapshotLifecycleManagement.ExecuteSnapshotLifecycleAsync(new ExecuteSnapshotLifecycleRequest("policy_id")))
-				;
-
-			await DELETE("/_slm/policy/policy_id")
-					.Fluent(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycle("policy_id"))
-					.Request(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycle(new DeleteSnapshotLifecycleRequest("policy_id")))
-					.FluentAsync(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycleAsync("policy_id"))
-					.RequestAsync(c => c.SnapshotLifecycleManagement.DeleteSnapshotLifecycleAsync(new DeleteSnapshotLifecycleRequest("policy_id")))
 				;
 		}
 	}
